feat(Data): move data management to Data.vue
